# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateYoutubeTranscriptionDto": {
        "properties": {
          "channelId": {
            "type": "string"
          },
          "customPrompt": {
            "maxLength": 500,
            "type": "string"
          },
          "generateAudio": {
            "type": "boolean"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "url",
          "channelId"
        ],
        "type": "object"
      },
...
      "UnauthorizedResponseDto": {
        "properties": {
          "message": {
            "example": "Unauthorized",
            "type": "string"
          },
          "statusCode": {
            "example": 401,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "type": "object"
      },
...
      "ValidationErrorResponseDto": {
        "properties": {
          "error": {
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "example": [
              "source_id must be a valid UUID"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "statusCode": {
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Meridiano API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/youtube/transcriptions": {
      "get": {
        "operationId": "YoutubeTranscriptionsController_listTranscriptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of YouTube transcriptions"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "List the authenticated user's YouTube transcriptions",
        "tags": [
          "YoutubeTranscriptions"
        ]
      },
      "post": {
        "operationId": "YoutubeTranscriptionsController_createTranscription",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateYoutubeTranscriptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Transcription created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Create a transcription for a YouTube video",
        "tags": [
          "YoutubeTranscriptions"
        ]
      }
    },
    "/api/youtube/transcriptions/{id}": {
      "delete": {
        "operationId": "YoutubeTranscriptionsController_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcription deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Delete a YouTube transcription",
        "tags": [
          "YoutubeTranscriptions"
        ]
      },
      "get": {
        "operationId": "YoutubeTranscriptionsController_getTranscription",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeAudio",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcription retrieved"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "YouTube transcription not found"
          }
        },
        "summary": "Get a YouTube transcription by id",
        "tags": [
          "YoutubeTranscriptions"
        ]
      }
    },
    "/api/youtube/transcriptions/{id}/audio": {
      "post": {
        "operationId": "YoutubeTranscriptionsController_generateAudio",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Audio generation job accepted"
          },
          "400": {
            "description": "Transcription has no content available for audio generation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "YouTube transcription not found"
          },
          "409": {
            "description": "Audio already exists or generation already in progress"
          }
        },
        "summary": "Generate audio for a YouTube transcription",
        "tags": [
          "YoutubeTranscriptions"
        ]
      }
    },
...
  },
  "servers": [],
...
}